# openapi.yaml
openapi: 3.0.3
info:
  title: Facturation.PRO API
  version: "1.0.0"
  description: API REST pour l'outil de facturation Facturation.PRO
  contact:
    email: support@facturation.pro
  license:
    name: Propriétaire
servers:
  - url: https://www.facturation.pro/
paths:
...
  /firms/{firmId}/customers.json:
    parameters:
      - name: firmId
        in: path
        required: true
        schema:
          type: integer
      - name: page
        in: query
        schema:
          type: integer
      - name: mode
        in: query
        schema:
          type: string
          enum: [all, company, individual, archived]
      - name: api_id
        in: query
        schema:
          type: integer
      - name: api_custom
        in: query
        schema:
          type: string
      - name: company
        in: query
        schema:
          type: string
      - name: last_name
        in: query
        schema:
          type: string
      - name: email
        in: query
        schema:
          type: string
      - name: category_id
        in: query
        schema:
          type: integer
      - name: with_sepa
        in: query
        schema:
          type: integer
      - name: account_code
        in: query
        schema:
          type: string
      - name: accounting_entry
        in: query
        schema:
          type: string
      - name: sort
        in: query
        schema:
          type: string
          enum: [asc, desc]
      - name: order
        in: query
        schema:
          type: string
          enum: [last_invoice, last_paid, created, updated]
    get:
      summary: Liste des clients
      operationId: listCustomers
      responses:
        '200':
          description: Liste des clients
          headers:
            X-Pagination:
              description: Informations de pagination
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    post:
      summary: Créer un client
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                company_name:
                  type: string
                individual:
                  type: boolean
                first_name:
                  type: string
                last_name:
                  type: string
                street:
                  type: string
                city:
                  type: string
                zip_code:
                  type: string
                country:
                  type: string
                phone:
                  type: string
                fax:
                  type: string
                email:
                  type: string
                mobile:
                  type: string
                website:
                  type: string
                notes:
                  type: string
                vat_number:
                  type: string
                siret:
                  type: string
                category_id:
                  type: integer
                account_code:
                  type: string
                accounting_entry:
                  type: string
                currency:
                  type: string
                language:
                  type: string
                pay_before:
                  type: string
                reference:
                  type: string
                field1:
                  type: string
                field2:
                  type: string
                field3:
                  type: string
                field4:
                  type: string
                field5:
                  type: string
      responses:
        '201':
          description: Client créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
...
  /firms/{firmId}/quotes.json:
    parameters:
      - name: firmId
        in: path
        required: true
        schema:
          type: integer
      - name: page
        in: query
        schema:
          type: integer
      - name: with_details
        in: query
        schema:
          type: integer
          enum: [0, 1]
      - name: api_id
        in: query
        schema:
          type: integer
      - name: api_custom
        in: query
        schema:
          type: string
      - name: quote_ref
        in: query
        schema:
          type: string
      - name: full_quote_ref
        in: query
        schema:
          type: string
      - name: customer_id
        in: query
        schema:
          type: integer
      - name: title
        in: query
        schema:
          type: string
      - name: company
        in: query
        schema:
          type: string
      - name: last_name
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
      - name: category_id
        in: query
        schema:
          type: integer
      - name: followup_id
        in: query
        schema:
          type: integer
      - name: accounting_entry
        in: query
        schema:
          type: string
      - name: user_id
        in: query
        schema:
          type: integer
      - name: period_start
        in: query
        schema:
          type: string
      - name: period_end
        in: query
        schema:
          type: string
      - name: sort
        in: query
        schema:
          type: string
          enum: [asc, desc]
      - name: order
        in: query
        schema:
          type: string
          enum: [customer, total, billed, created, updated]
    get:
      summary: Liste des devis
      operationId: listQuotes
      responses:
        '200':
          description: Liste des devis
          headers:
            X-Pagination:
              description: Informations de pagination
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Quote'
    post:
      summary: Créer un devis
      operationId: createQuote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - customer_id
                - invoiced_on
                - language
                - items
              properties:
                customer_id:
                  type: integer
                currency:
                  type: string
                invoiced_on:
                  type: string
                  format: date
                term_on:
                  type: string
                  format: date
                language:
                  type: string
                pay_before:
                  type: string
                penalty:
                  type: number
                title:
                  type: string
                rebate_percentage:
                  type: number
                vat_exemption:
                  type: string
                information:
                  type: string
                category_id:
                  type: integer
                followup_id:
                  type: integer
                user_id:
                  type: integer
                accounting_entry:
                  type: string
                delivery_on:
                  type: string
                  format: date
                purchase_number:
                  type: string
                internal_note:
                  type: string
                activity_title:
                  type: string
                service_personne:
                  type: boolean
                tax_title:
                  type: string
                tax_percent:
                  type: number
                precompte:
                  type: boolean
                vat_country:
                  type: string
                api_id:
                  type: integer
                api_custom:
                  type: string
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/QuoteItem'
      responses:
        '201':
          description: Devis créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
...
  /firms/{firmId}/quotes/{id}/invoice.json:
    parameters:
      - name: firmId
        in: path
        required: true
        schema:
          type: integer
      - name: id
        in: path
        required: true
        schema:
          type: integer
    post:
      summary: Convertir un devis en facture
      operationId: convertQuoteToInvoice
      responses:
        '201':
          description: Facture créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
...
  /firms/{firmId}/invoices.json:
    parameters:
      - name: firmId
        in: path
        required: true
        schema:
          type: integer
      - name: page
        in: query
        schema:
          type: integer
      - name: with_details
        in: query
        schema:
          type: integer
          enum: [0, 1]
      - name: with_settlements
        in: query
        schema:
          type: integer
          enum: [0, 1]
      - name: api_id
        in: query
        schema:
          type: integer
      - name: api_custom
        in: query
        schema:
          type: string
      - name: invoice_ref
        in: query
        schema:
          type: string
      - name: full_invoice_ref
        in: query
        schema:
          type: string
      - name: payment_ref
        in: query
        schema:
          type: string
      - name: title
        in: query
        schema:
          type: string
      - name: customer_id
        in: query
        schema:
          type: integer
      - name: company
        in: query
        schema:
          type: string
      - name: last_name
        in: query
        schema:
          type: string
      - name: bill_type
        in: query
        schema:
          type: string
          enum: [paid, unpaid, term, invoice, external, refund, down_payment, draft, notes, nova]
      - name: category_id
        in: query
        schema:
          type: integer
      - name: followup_id
        in: query
        schema:
          type: integer
      - name: accounting_entry
        in: query
        schema:
          type: string
      - name: user_id
        in: query
        schema:
          type: integer
      - name: period_start
        in: query
        schema:
          type: string
      - name: period_end
        in: query
        schema:
          type: string
      - name: period_type
        in: query
        schema:
          type: string
      - name: sort
        in: query
        schema:
          type: string
          enum: [asc, desc]
      - name: order
        in: query
        schema:
          type: string
          enum: [customer, paid, total, billed, term, created, updated]
    get:
      summary: Liste des factures
      operationId: listInvoices
      responses:
        '200':
          description: Liste des factures
          headers:
            X-Pagination:
              description: Informations de pagination
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
    post:
      summary: Créer une facture
      operationId: createInvoice
      parameters:
        - name: type_doc
          in: query
          schema:
            type: string
        - name: external
          in: query
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - customer_id
                - invoiced_on
                - language
                - items
              properties:
                customer_id:
                  type: integer
                currency:
                  type: string
                invoiced_on:
                  type: string
                  format: date
                language:
                  type: string
                pay_before:
                  type: string
                penalty:
                  type: number
                title:
                  type: string
                rebate_percentage:
                  type: number
                vat_exemption:
                  type: string
                information:
                  type: string
                category_id:
                  type: integer
                followup_id:
                  type: integer
                user_id:
                  type: integer
                accounting_entry:
                  type: string
                delivery_on:
                  type: string
                  format: date
                purchase_number:
                  type: string
                internal_note:
                  type: string
                activity_title:
                  type: string
                service_personne:
                  type: boolean
                tax_title:
                  type: string
                tax_percent:
                  type: number
                precompte:
                  type: boolean
                vat_country:
                  type: string
                external_ref:
                  type: string
                api_id:
                  type: integer
                api_custom:
                  type: string
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/InvoiceItem'
      responses:
        '201':
          description: Facture créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
  /firms/{firmId}/invoices/{id}.json:
    parameters:
      - name: firmId
        in: path
        required: true
        schema:
          type: integer
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: type_doc
        in: query
        schema:
          type: string
        description: final ou draft
    get:
      summary: Détails d'une facture
      operationId: getInvoice
      responses:
        '200':
          description: Détails de la facture
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
    patch:
      summary: Modifier une facture
      operationId: updateInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customer_id:
                  type: integer
                currency:
                  type: string
                invoiced_on:
                  type: string
                  format: date
                language:
                  type: string
                pay_before:
                  type: string
                penalty:
                  type: number
                title:
                  type: string
                rebate_percentage:
                  type: number
                vat_exemption:
                  type: string
                information:
                  type: string
                category_id:
                  type: integer
                followup_id:
                  type: integer
                user_id:
                  type: integer
                accounting_entry:
                  type: string
                delivery_on:
                  type: string
                  format: date
                purchase_number:
                  type: string
                internal_note:
                  type: string
                activity_title:
                  type: string
                service_personne:
                  type: boolean
                tax_title:
                  type: string
                tax_percent:
                  type: number
                precompte:
                  type: boolean
                vat_country:
                  type: string
                payment_mode:
                  type: integer
                paid_on:
                  type: string
                  format: date
                payment_ref:
                  type: string
                api_id:
                  type: integer
                api_custom:
                  type: string
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/InvoiceItem'
      responses:
        '200':
          description: Facture modifiée
    delete:
      summary: Supprimer une facture
      operationId: deleteInvoice
      responses:
        '200':
          description: Facture supprimée
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
          description: Identifiant
        company_name:
          type: string
          description: Société
        civility:
          type: string
          description: Civilité
        first_name:
          type: string
          description: Prénom
        last_name:
          type: string
          description: Nom
        short_name:
          type: string
          description: Nom mnémotechnique
        street:
          type: string
          description: Adresse de facturation
        city:
          type: string
          description: Ville
        zip_code:
          type: string
          description: Code postal
        country:
          type: string
          description: Pays (Code ISO à 2 lettres du pays, en majuscules)
        phone:
          type: string
          description: Téléphone
        fax:
          type: string
          description: Fax
        email:
          type: string
          description: Email
        mobile:
          type: string
          description: Mobile
        website:
          type: string
          description: Site web
        notes:
          type: string
          description: Notes internes
        vat_number:
          type: string
          description: N° TVA
        siret:
          type: string
          description: Siret
        category_id:
          type: integer
          description: Catégorie
        account_code:
          type: string
          description: Compte client
        accounting_entry:
          type: string
          description: Compte d'imputation
        created_at:
          type: string
          format: date-time
          description: Créé le
        updated_at:
          type: string
          format: date-time
          description: Modifié le
        sepa_iban:
          type: string
          description: IBAN
        sepa_bic:
          type: string
          description: BIC
        sepa_rum:
          type: string
          description: RUM
        sepa_signature_on:
          type: string
          format: date
          description: Date de signature du mandat SEPA
        api_id:
          type: integer
          description: API - Référence numérique libre
        api_custom:
          type: string
          description: API - Texte libre
        individual:
          type: boolean
          description: Particulier
        reference:
          type: string
          description: Référence client
        penalty:
          type: number
          description: Pénalités de retard
        discount:
          type: number
          description: Taux d'escompte
        pay_before:
          type: string
          description: Délai de paiement
        validity:
          type: integer
          description: Durée de validité d'un devis
        last_invoiced_on:
          type: string
          format: date
          description: Dernière facture le
        last_paid_on:
          type: string
          format: date
          description: Dernier règlement le
        currency:
          type: string
          description: Devise
        language:
          type: string
          description: Langue des PDF
        default_vat:
          type: number
          description: Taux de TVA
        vat_country:
          type: string
          description: Code pays des taux TVA à appliquer
        vat_exemption:
          type: string
          description: Motif d'exonération de TVA
        invoice_email:
          type: string
          description: Email facturation
        field1:
          type: string
          description: Champ libre 1
        field2:
          type: string
          description: Champ libre 2
        field3:
          type: string
          description: Champ libre 3
        field4:
          type: string
          description: Champ libre 4
        field5:
          type: string
          description: Champ libre 5
        soft_deleted:
          type: boolean
          description: Enregistrement marqué comme à supprimer
        hard_delete_on:
          type: string
          format: date
          description: Date définitive de la suppression
...
    QuoteItem:
      type: object
      properties:
        id:
          type: integer
          description: Identifiant
        quantity:
          type: number
          description: Quantité
        measure:
          type: string
          description: Unité de mesure
        title:
          type: string
          description: Libellé
        unit_price:
          type: number
          description: Prix unitaire HT
        vat:
          type: number
          description: Taux de TVA
        product_id:
          type: integer
          description: Identifiant du produit
        position:
          type: integer
          description: Ligne n°
        total:
          type: number
          description: Total HT
        optional:
          type: boolean
          description: En option
        style:
          type: string
          description: Style de formattage
        nature:
          type: integer
          description: Type
        sap_nature:
          type: integer
          description: Code activité
...